fixup MDEV-27277 Add a warning when max_sort_length is reached Step#1
fulltest2:

/home/buildbot/buildbot/build/mariadb-11.7.0/strings/ctype-mb.c: In function ‘my_strnxfrm_mb_internal’:
/home/buildbot/buildbot/build/mariadb-11.7.0/strings/ctype-mb.c:504:47: error: signed and unsigned type in conditional expression [-Werror=sign-compare]
       size_t len= (dst + chlen <= de) ? chlen : de - dst;

/home/buildbot/buildbot/build/mariadb-11.7.0/strings/ctype-utf8.c: In function ‘my_strnxfrm_unicode_full_bin’:
/home/buildbot/buildbot/build/mariadb-11.7.0/strings/ctype-utf8.c:318:20: error: comparison between signed and unsigned integer expressions [-Werror=sign-compare]
     if ((de - dst) < nweights * 3)

+++ strings/ctype-mb.c
@@ -501,7 +501,7 @@ my_strnxfrm_ret_t my_strnxfrm_mb_internal(CHARSET_INFO *cs,
     else
     {
       /* Multi-byte character */
-      size_t len= (dst + chlen <= de) ? chlen : de - dst;
+      size_t len= (dst + chlen <= de) ? (size_t) chlen : (size_t) (de - dst);
       if (dst + chlen > de)
         warnings|= MY_STRNXFRM_TRUNCATED_WEIGHT_REAL_CHAR;
       memcpy(dst, src, len);
